Updating with insanely easy docker script.
Now, there is no excuse for people claiming the instructions don't work,
since there is a 100% reproducable docker file which just _does_ the install
for you. There is now a super easy script which will quickly bind a mount
point for the file, pass it to apkid and get the result. \o/